build(commitlint): Add commitlint to development environment
Also add/configure husky.  My feelings on husky are a little mixed as it
basically takes over all git hooks and git hooks aren't exactly
difficult to manage so it's hard to say it's definitely worth it.  But,
it _is_ nice to have git hooks set up automatically as part of `yarn
install`.